Moved SISpinbox to its own file
